Move isSupported method to module
